pyproject.toml
[project]
name = "smoke-fire-detection"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "matplotlib>=3.10.1",
    "numpy<2.0.0",
    "opencv-python>=4.11.0.86",
    "pandas>=2.2.3",
    "pillow>=11.1.0",
    "scikeras>=0.2.0",
    "scikit-learn>=1.6.1",
    "seaborn>=0.13.2",
    "tensorflow==2.15.0",
    "tqdm>=4.67.1",
    "ultralytics>=8.3.107",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["."]
